+++ server/src/modules/media/dto/update-own-media.dto.ts
@@ -1,10 +1,5 @@
 import { ApiPropertyOptional } from '@nestjs/swagger';
-import {
-  IsEnum,
-  IsNumber,
-  IsOptional,
-  IsString,
-} from 'class-validator';
+import { IsEnum, IsNumber, IsOptional, IsString } from 'class-validator';
 import { MediaOwnerType } from '../../../entities/media.entity';
 
 export class UpdateOwnMediaDto {